+++ packages/hippobase/hippobase_core_models/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.2
+
+* Allow list queries to omit pagination and sorting.
+* Allow pagination limits to be omitted, with offsets defaulting to zero.
+
 ## 0.1.1
 
 * Require `json_schema` and `json_schema_gen` 0.1.1.